Metal, Dieletric, subpix-rendering and fresnel fix
Barring the addition of Schlick reflectance on Metal, I believe it's
mostly on par with the C++ version. The hittable list wasn't implemented
as it offered no sensible purpose and acts mostly as a temporary vector
on the C++ side.

Now comes the !!fun!! part where I'll try to make the code more
idiomatic.